f1/analysis_engine/requirements.txt
fastf1==0.6.10
google-generativeai==0.3.0
pandas==2.1.0
numpy==1.24.0
requests==2.31.0
python-dotenv==1.0.0
